cinfig refactoring
