+++ node/src/main/resources/reference.conf
@@ -10,8 +10,6 @@ endpoints {
 }
 
 swarm {
-  // use the Swarm or run local precompiled code
-  enabled = true,
   // the address to the Swarm gateway
   host = "http://localhost:8500"
 }
